[Change Log] History Issue Fixed

+++ app/src/main/java/com/gox/partner/models/CheckRequestModel.kt
@@ -216,7 +216,7 @@ data class Invoice(
 ) : Serializable
 
 data class Payment(
-        var card: Int? = 0,
+        var card: Double? = 0.0,
         var cash: Double? = 0.0,
         var commision: Double? = 0.0,
         var commision_percent: Double? = 0.0,
